fix animations on android

+++ assets/engine/shaders/android_geometry.vert
@@ -6,9 +6,8 @@ layout(location = 0) in vec3 aPos;
 layout(location = 1) in vec3 aNormal;
 layout(location = 2) in vec2 aTexCoords;
 
-// int because opengl has beef with ints istg
-layout(location = 3) in vec4 aBoneIndices;       // the indices of the bones that affect this particular vertex
-layout(location = 4) in vec4 aBoneWeights;        // how much effect does this bone have on this vertex.
+layout(location = 3) in ivec4 aBoneIndices;      // the indices of the bones that affect this particular vertex
+layout(location = 4) in vec4 aBoneWeights;       // how much effect does this bone have on this vertex.
 // in this case, there are a max of 4 bones that can affect a vertex. all 4 bones must add up to 1
 
 // normal map stuff
@@ -36,10 +35,10 @@ void main() {
     vec3 localNormal;
     
     if (u_Animated > 0.5) {
-        mat4 skin = u_BoneMatrices[int(aBoneIndices.x)] * aBoneWeights.x
-                  + u_BoneMatrices[int(aBoneIndices.y)] * aBoneWeights.y
-                  + u_BoneMatrices[int(aBoneIndices.z)] * aBoneWeights.z
-                  + u_BoneMatrices[int(aBoneIndices.w)] * aBoneWeights.w;
+        mat4 skin = u_BoneMatrices[aBoneIndices.x] * aBoneWeights.x
+                  + u_BoneMatrices[aBoneIndices.y] * aBoneWeights.y
+                  + u_BoneMatrices[aBoneIndices.z] * aBoneWeights.z
+                  + u_BoneMatrices[aBoneIndices.w] * aBoneWeights.w;
 
         // vec4 hardcodedWeights = vec4(0.25, 0.25, 0.25, 0.25);
         // ivec4 hardcodedIndices = ivec4(0, 0, 0, 0);  // All bone 0
